+++ jest.config.ts
@@ -1,0 +1,4 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const { getJestProjects } = require('@nrwl/jest');
+
+export default { projects: getJestProjects() };